Move common pipeline types into subfolders
Small bit of tidying and renaming to organise `PublishingEventPipeline`
source files better.
