+++ lib/publishing_event_pipeline/events/unpublish.rb
@@ -1,0 +1,16 @@
+module PublishingEventPipeline
+  module Events
+    class Unpublish
+      attr_reader :content_id, :payload_version
+
+      def initialize(content_id, payload_version:)
+        @content_id = content_id
+        @payload_version = payload_version
+      end
+
+      def synchronize_to(repository)
+        repository.delete(content_id, payload_version:)
+      end
+    end
+  end
+end
