--- lib/publishing_event_pipeline/metadata_extractor.rb
@@ -1,9 +1,0 @@
-module PublishingEventPipeline
-  class MetadataExtractor
-    def call(message_hash)
-      {
-        base_path: message_hash.fetch("base_path"),
-      }
-    end
-  end
-end
